// src/app/model/policy.enum.ts
export enum PolicyEnum {
  PARTNER_SAMSUNG_PHILLIPINES_RETAIL = 'PHSAMSUOEMNA01',
  PARTNER_SAMSUNG_INDONESIA_RETAIL = 'IDSAMSUOEMNA01',
  PARTNER_SAMSUNG_INDONESIA_CF = 'IDSAMSUOEMCF01',
  PARTNER_SAMSUNG_PHILLIPINES_CF = 'PHSAMSUOEMCF01',
  PARTNER_SAMSUNG_VIETNAM_RETAIL = 'VNSAMSUOEMNA01',
  PARTNER_SAMSUNG_VIETNAM_RETAIL_GRAB = 'VNSAMSUOEMNA02',
  PARTNER_HOMECREDIT_VIETNAM = 'VNHOMCRCFINA01',
  PARTNER_DIGI_MALAYSIA = 'MYDIGIMTELNA01',
  EDIT_LOCK_MESSAGE = 'Editing of policy'
}